Refactor day 03
No number touches two parts, so it's safe to collect a list of adjacent
numbers for each symbol instead of collecting all symbol-number pairs.
This lets us calculate the gear ratios more conveniently, because we
don't have to do the extra step of collecting all adjacent numbers for
each coordinate before doing the product. We also don't need to store
the coordinates anymore.