Day 16: refactor out unnecessary GridDict
As we're reading in the whole grid anyway, there's no reason to use a
data structure that returns a default value for unknown keys. In fact
using a GridDict only increases the possibility of coding errors--in
this puzzle we're supposed to stay wholly inside the grid at all times,
but a GridDict would happily return values for coordinates outside the
grid, whereas with a normal dict will raise a key error.